+++ test/e2e.ts
@@ -6,7 +6,8 @@ import { promises as fs } from "node:fs";
 import { dirname, join } from "node:path";
 import process from "node:process";
 import { fileURLToPath } from "node:url";
-import { createLogger } from "@schemasset/utils";
+// eslint-disable-next-line antfu/no-import-dist
+import { createLogger } from "../packages/utils/dist/index.js";
 
 // Create a logger for testing
 const logger = createLogger({ prefix: "test" });