+++ CHANGELOG.md
@@ -1,6 +1,36 @@
 # Changelog
 
 
+## v0.0.3
+
+
+### 🚀 Enhancements
+
+- Parser ([#2](https://github.com/mates-system/schemasset/pull/2))
+- Checker ([#3](https://github.com/mates-system/schemasset/pull/3))
+- Brand ([#5](https://github.com/mates-system/schemasset/pull/5))
+- Nuxt Module ([#6](https://github.com/mates-system/schemasset/pull/6))
+
+### 🩹 Fixes
+
+- Checker algorithm ([#4](https://github.com/mates-system/schemasset/pull/4))
+- Using nuxt module build ([096a560](https://github.com/mates-system/schemasset/commit/096a560))
+
+### 🏡 Chore
+
+- Configure lint ([#1](https://github.com/mates-system/schemasset/pull/1))
+- Tsconfig ([f5e0d2c](https://github.com/mates-system/schemasset/commit/f5e0d2c))
+- Example code ([e6851f4](https://github.com/mates-system/schemasset/commit/e6851f4))
+- Npm config ([#8](https://github.com/mates-system/schemasset/pull/8))
+
+### 🤖 CI
+
+- Configure ([#7](https://github.com/mates-system/schemasset/pull/7))
+
+### ❤️ Contributors
+
+- [ubugeeei](https://github.com/ubugeeei)
+
 ## v0.0.2